fix folder not exist

+++ build.js
@@ -1,5 +1,8 @@
 const fs = require('fs')
-fs.rmdirSync('dist', { recursive: true })
+
+if (fs.existsSync('dist')) {
+    fs.rmdirSync('dist', { recursive: true })
+}
 fs.mkdirSync('dist')
 fs.copyFileSync('lib/torch.js', 'dist/torch.js')
 fs.copyFileSync('lib/torch.d.ts', 'dist/torch.d.ts')
